+++ Packages/rtrt/Core/TileImageMaterial.cc
@@ -32,7 +32,7 @@ PersistentTypeID TileImageMaterial::type_id("TileImageMaterial", "Material",
 
 TileImageMaterial::TileImageMaterial( int /*oldstyle*/, const string &texfile, 
 				      double Kd,
-				      const Color& specular, double specpow,
+				      const Color& specular, int specpow,
 				      double refl, bool flipped/*=false*/ ) :
   ImageMaterial(1,texfile,ImageMaterial::Tile,ImageMaterial::Tile,
 		Kd,specular,specpow,refl,flipped)
@@ -41,7 +41,7 @@ TileImageMaterial::TileImageMaterial( int /*oldstyle*/, const string &texfile,
 
 TileImageMaterial::TileImageMaterial( const string &texfile, 
 				      double Kd,
-				      const Color& specular, double specpow,
+				      const Color& specular, int specpow,
 				      double refl, bool flipped /*= false*/ ) :
   ImageMaterial(texfile,ImageMaterial::Tile,ImageMaterial::Tile,
 		specular,Kd,specpow,refl,flipped)
@@ -50,7 +50,7 @@ TileImageMaterial::TileImageMaterial( const string &texfile,
 
 TileImageMaterial::TileImageMaterial(const string &texfile, 
                                      double Kd,
-                                     const Color& specular, double specpow,
+                                     const Color& specular, int specpow,
                                      double refl,  double transp, 
                                      bool flipped /*=false*/) :
   ImageMaterial(texfile,ImageMaterial::Tile,ImageMaterial::Tile,
